# Conflicts:
#	Application/Application.vcxproj
#	Application/Application.vcxproj.filters
#	Application/Source/Application.cpp
#	Application/Source/Application.h
#	Application/Source/MGPlayer.cpp
#	Application/Source/MainMenu.cpp
#	Application/Source/MainScene.cpp

+++ Application/Source/Inventory.cpp
@@ -107,7 +107,7 @@ void Inventory::Update(double dt)
 	{
 		for (size_t i = 0; i < Inventory_.size(); i++)
 		{
-			if (Inventory_[i]->hitBox->isPointInsideAABB(Position(cur_x, Application::getWindowHeight() - cur_y, 0.f)))
+			if (Inventory_[i]->hitBox->isPointInsideAABB(Position(cur_x, Application::getWindowHeight() - (float)cur_y, 0.f)))
 			{
 				Inventory::getItem(i);
 				break;
